requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
pandas
numpy
datetime 
joblib
scikit-learn==1.7.1
fastapi-utils>=0.2.1
pika>=1.2.0
motor==2.5.1
pymongo==3.12.3
aio-pika==6.8.2
uvicorn[standard]
websockets
